pyproject.toml
[project]
name = "co_genticai"
version = "0.1.0"
description = "A Collaborative Agentic AI System"
authors = [
    {name = "ems.ameen@gmail.com"}
]
packages = [{include = "co_genticai", from = "src"}]
license = {text = "Apatche 2.0"}
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.13"
streamlit = "^1.32.0"
python-dotenv = "^1.0.1"
google-generativeai = "^0.3.2"
loguru = "^0.7.2"
pandas = "^2.2"
pinecone-client = "^3.0.0"
sentence-transformers = "^4.1.0"

# LangChain Core and Plugins (Pinned)
langchain = ">=0.1.14"
langchain-community = ">=0.0.27"
langchain-openai = ">=0.0.8"
langchain-huggingface = ">=0.0.5"
langchain-pinecone = ">=0.0.3"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.2"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
